Package style folder structure; setup.py
In order to make this project something that can be installed simply and
consistently, I made a few slightly large structural changes:
  - moved all the python files for this project into the directory xsboa
  - created a setup.py file
  - added __init__.py to the xsboa directory

The setup.py file allows the project to be installed with the command
  python setup.py install
Doing this compiles all the necessary modules so the submodules and subfunctions
can be imported from any directory like a normal python package

The __init__.py file in the xsboa directory essentially tells the import
command what modules and functions are valid to be imported. On this
branch, only messages is present, but down the road modules like scraper
and any plotting and preparation files will be included here as well.

Signed-off-by: drewejohnson <1drew.e.johnson@gmail.com>