Extract variables from parameter block
Added keyword var that, if detected in the parameter block, extracts the
desired keywords and variables to be scraped from the SERPENT _res.m
files. This does not have to exist for creating and executing the SERPENT files,
but processing the outputs without this list doesn't make a ton of sense.

Updated test_readparam.py
  - updated location of output tuple (had to include 2 new lines for var)
  - new test to validate correct extraction of variables

Signed-off-by: drewejohnson <1drew.e.johnson@gmail.com>