Updated res_scraper to work for cases without burnup
It was noted that, if the _res.m file did not have any burnup points, then the res_scraper function would not be adequate. Support for the burnlist and burntype arguments to be None, indicating that burnup should not be used in the extracting of the values. Testing has been updated accordingly.

Signed-off-by: drewejohnson <1drew.e.johnson@gmail.com>