Burnup strings no longer pulled from read-param as unique list; added to branch dictionary (#6)
* Updated readparam - burnup string in branch dictionary; Cleaned docstring

The burnup string, if specified, is added to the corresponding branch, rather than an integer pointing to an index in another vector. That's silly. Updated the testing accordingly.

exe_str should be setup to allow for use of the .format() method, where the only argument is the filename i.e. ./sss2 {0} > {0}.o

Cleaned up some docstrings to make things look a little cleaner/comply with some PEP8 standards. These changes should help with documentation later.

Signed-off-by: drewejohnson <1drew.e.johnson@gmail.com>

* Updated readparam - burnup string in branch dictionary; Cleaned docstring

The burnup string, if specified, is added to the corresponding branch, rather than an integer pointing to an index in another vector. That's silly. Updated the testing accordingly.

exe_str should be setup to allow for use of the .format() method, where the only argument is the filename i.e. ./sss2 {0} > {0}.o

Cleaned up some docstrings to make things look a little cleaner/comply with some PEP8 standards. These changes should help with documentation later.

Signed-off-by: drewejohnson <1drew.e.johnson@gmail.com>

Final - burnup strings added to specific branches, not unique list

Signed-off-by: drewejohnson <1drew.e.johnson@gmail.com>

Updated to 0.1.9

Signed-off-by: drewejohnson <1drew.e.johnson@gmail.com>

* Updated readparam - branching covers perturbation space

Branching dictionary returned from readparam() now describes the name of the branch, material being perturbed, perturbation type, and then a vector of perturbation values. This will allow all combintations of nominal and perturbed conditions to be branched

Signed-off-by: drewejohnson <1drew.e.johnson@gmail.com>